Problems with the icons
So the icons that the API gives me didnt work, and
even if the did, they are not suited for my design.

The issue is that the API retrieves a lot of different
weather names, so i have to group them into categories,
and then display an icon for each category.

I couldnt find a logic that is useful and not horrible,
at least yet.