Fixed icons logic and added DOM card
First, the icons problem is resolved. I take all the possible
values and put it in different categories (arrays). Then, pass
them through the iconSelector function, which has a pretty
simple logic.

Regarding to the DOM card, i created the first card which is
displayed in the principal part of the main content of the
user interface.