Added another function to create cards
Now the program have two functions to fill the cards:
the first one is currentHourCard and is going to be
used by current and hours object.
The second one is fullDayCard and is going to be used
by history and forecast days.
This is because this objects share almost the same
properties, but im probably going to look over this
logic in the next days.

Then i added some properties (isYesterday and isCelsius)
in the objects that i create (with the disaggregated API
data) that i think are going to help me to display some
things and not others, depending on the user preference
(celsius or faranheint) or if the day has already passed
or not (rain chances % or mm).