Changed how to get the structured data
It was done through 3 different functions: one for
forecast data, one for history data, and one for
current and location data.
Now is just one function (getStructuredData) and
i call from there, the different functions that display
the cards (one for current, other for history and
forecast days, and the last one for every hour).
This is the structure of the code.. for now. I may
change it in the next days, to give the app different
ui options (like click on a forecast day, like tomorrow,
and display its data) or just because i found a better
logic to apply.

Also, fixed a little bug in the iconSelector function,
the default value with the or (||) didnt work because
the value zero (0) breaks it. So i use a ternary operator
instead.