Changed the main function getData and added a file
Now in getData function, i call 2 major functions:
triggerStructuration and triggerDisplay (these 2 for now).
The former handles the structuring of the fetched data,
using the functions of 2 files: objects.js and class.js.
The latter handles the display of that data, using the
functions of other 2 files: dom.js and (the new one)
display.js.