Simplified the setting of the data in the elements
Now there is just one function that is in charge of that,
setCardData(), which is in the don.js file. It replace
fullDayCard and currentHourCard functions.

Also change the CSS syntax of the elements not displayed,
making it more simple (with the :not() pseudoclass).

Finally, started with the carousel to display the hours
element in the interface (not finished yet).