Create objects now, and added the date property
In the objects.js file, now the functions getForecastData
and getHistoryData, creates objects, for each one: day,
yesterday, tomorrow and lastDay.
This way is easier to call with accesors and not with
destructuring (before i have to do that, because i was
creating arrays and not objects).

Also, added the date property inside the constructors,
so now i need one less argument when i call the display
functions.