Change the date for the history fetch
I do it manually, because i didnt make the function yet.
Probably im going to import a function of a library that
work with dates and calendars.
If i dont change the history date, when some time passes by, the
API throw an error. Thats why i have to change it.