Changed carousel and auxiliary function
First, changed the carousel becuase having both
hour containers overflowing, make the layout stretch
a lot, even using the overflow:hidden property.
So used an animation to kind of replicate the
movement (not perfect at all, but the best i could).
Also changed the carousel element in the auxiliaries
file for the easyCar() function, to work with the
new conditions of the carousel.

Also, had to modify the iconSelector funciton, in the
auxiliaries file, which failed becuase the API retrieve
the textWeather property with a typo, an extra space after
the word. So i take a little bit of extra care for the
strings, adding the trim() and toLowerCase() methods,
just in case.