Changed all the layouts of the cards or sections
Changed the principal section, the extra section and
the near section.
All changed for grid with auto-fit, which allows the
elements to respond to the viewport width, making the
overall layout much more responsive.
Also added two media queries, one to change the body
layout and put the aside section below the other sections
when the width of the viewport is 600px or less, and the
other to put a min-height in the aside section when the
viewport is 600px or less.