Add media query to change principal card
The media query has a width of 990px, just when the
principal card change its layout to a stack.
This allow me to add a margin to the place element
when the viewport is less than that, and to add a
negative margin to the temp element when the viewport
is bigger than that.
This way the interface looks better.

Also, added a margin-bottom to the aside elements
(each hour card) when the media query of the 600px
width kicks off, again to make the interface looks
better.