Add drop down menu and an option in auxiliaries
The drop down menu (hamburguer button) is beside the
input. The options allow the user to change the color,
temperature scale and MAYBE the languaje.
At the moment i only done the interface of the menu,
it lacks the functionality.

Also, added a new option in the iconsObj of the
auxiliaries file. This new option is not in the
documents of the API, but it retrieve that value once
so i added it.