Changes in getData() and triggerDataDisplay()
In getData(), added management of errors. If the
status of the fetch is not "ok", the function throw an
error to properly catch it. Then, separated the other
functions that use to be inside this one:
triggerDataStructuration() and triggerDataDisplay()
are now inside dataHandler(). I dont know if im going
to keep it this way tho. For now, is fine.

About the changes in triggerDataDisplay(), is now prepared
to display data of different days, that could be yesterday,
today, tomorrow or lastDay. But i have to figure out the
way to make this function to listen to the respective events.