App can change the data day display
Now the user can click on a day (yesterday, today,
tomorrow or lastDay) on the "near" section, and
the interface will display that data.
To make it work i have to fetch the data again.
Maybe i can do that more efficiently, because the
data is already fetched, so no need to do that again.
But for now, is good enough.