Changes to extra card and dom file
Added the thermic sensation to the extra card (only
when its the current weather beacuse i dont have the
data for the others days).

Also build a different function for each card in the
dom file (principal, near, extra and hour). Its much more
repetitive, but maybe its clearer to read. If i keep
changing the things that the cards display when the user
change the day, i believe its going to make the code cleaner.