Added storeData() and retrieveData()
this are the functions that the app was lacking.
Now i dont need to fetch the data from the API
every single time the user make a change in the
interface.
Now it just fetch data when the user search for
another place.
This was the more efficient way i was looking for.
How i didnt think about this earlier? So obvious
haha.