Added handle error functionality
Three functions were created: createErrorCard()
in the dom file, renderError() and
handleErrorClass() in the display file.
The first function creates the elements, and
renderError() render that card when the input of
the user dont match any valid place. Its located
into the getData() function, where that function
catches the error.
handleErrorClass() is just an auxiliary function
that manage the classes used to render the
error card and layout in the interface.