Clean and organize the code.
put comments in all the functions (may be too much)
and added a new file, interactivity.js. This file
contains the functions that allow the user to
interact with the app, like getInputValue(),
showMenu(), changeTemp(), changeTheme(), etc.

Also started to clean the CSS file, deleting
redundancy and organizing it a little bit better.