Change the media query to 900px
At that point the layout change, and put the aside
element at the bottom.

Also changed the width of the image of the aside
card for a max(), which solves the problem of it
going too small. Same thing goes for the width of
the sub p, which has the weather description,
allowing to go wider if needed.