changed the recently added functionality and more
the functionality to determine if is currently night
or day, to choose the correct icon to display, is
now in the triggerDataDisplay() function.

Also, added some things in the CSS, to make it
more responsive. Most of all, paddings and margins
that change with the viewport height and width
(nothing crazy tho) in the body and other sections
of the web structure.

The margin-inline of the aside, with the responsive
fonts (with clamps) where the tougher to get, but
is very exciting to work with these functions.

I think the project is finished, at least for now.