Fixed one bug and one problem with text
The bug was in the recently added function to
determine if currently is day or night in the
place that the user search for. The problem was
that the sunny and clear days didnt share the same
weatherText property, like the others did.
So i have to add a function called
changeWeatherText() to swap them if needed. It is
inside the isDayOrNight() function in the display
file.

The other problem was that some of the weather
text was just too large. For that i created
the function stringLength() that get the text, its
length and then sort all of them. Then add a
function stringCut() with just a switch that take
care of the larger texts and repalce them for
other shorter. This stringCut function is in the
dom file, witihin the functions that set the data.

Also, i changed the font-size in the CSS for the
hour containers, to make it easier for the weather
text to fit in.